Fix startup races, move to separate .cpp
