+++ src/util/BatchExecutor.cpp
@@ -19,8 +19,6 @@
 namespace stellar
 {
 
-const size_t WORKER_STACK_BYTES = 1 << 23; // 8 MiB
-
 namespace
 {
 struct CpuPinning
@@ -162,8 +160,9 @@ BatchExecutor::ensureWorkers(size_t count)
     {
         size_t index = mWorkers.size();
         uint64_t batchId = mBatchId;
-        mWorkers.emplace_back(WORKER_STACK_BYTES,
-            [this, index, batchId]() { workerLoop(index, batchId); });
+        mWorkers.emplace_back(WORKER_STACK_BYTES, [this, index, batchId]() {
+            workerLoop(index, batchId);
+        });
         pinWorker(index);
     }
 }
