+++ src/util/StackThread.h
@@ -25,14 +25,12 @@
 
 #include <algorithm>
 #include <cstddef>
+#include <cstdint>
 #include <cstring>
-#include <exception>
 #include <functional>
 #include <memory>
 #include <ostream>
 #include <string>
-#include <system_error>
-#include <thread>
 #include <tuple>
 #include <type_traits>
 #include <utility>
@@ -45,16 +43,8 @@
 #define NOMINMAX
 #endif
 #include <windows.h>
-// windows.h first, then:
-#include <climits>
-#include <process.h>
-#ifndef STACK_SIZE_PARAM_IS_A_RESERVATION
-#define STACK_SIZE_PARAM_IS_A_RESERVATION 0x00010000
-#endif
 #else
-#include <limits.h>
 #include <pthread.h>
-#include <unistd.h>
 #endif
 
 namespace stellar
@@ -86,103 +76,17 @@ template <class Fn> struct Payload final : PayloadBase
     }
 };
 
-inline void
-setCurrentThreadName(std::string const& name)
-{
-    if (name.empty())
-    {
-        return;
-    }
-#if defined(_WIN32)
-    // SetThreadDescription is Windows 10 1607+; resolve dynamically so the
-    // binary still loads on older systems.
-    using SetDescFn = HRESULT(WINAPI*)(HANDLE, PCWSTR);
-    static SetDescFn const setDesc =
-        reinterpret_cast<SetDescFn>(reinterpret_cast<void*>(::GetProcAddress(
-            ::GetModuleHandleW(L"kernel32.dll"), "SetThreadDescription")));
-    if (setDesc != nullptr)
-    {
-        // Thread names are ASCII in practice; widen naively.
-        std::wstring wide(name.begin(), name.end());
-        setDesc(::GetCurrentThread(), wide.c_str());
-    }
-#elif defined(__APPLE__)
-    // Hard limit of 64 bytes including the NUL; longer names fail with ERANGE.
-    std::string const truncated = name.substr(0, 63);
-    ::pthread_setname_np(truncated.c_str());
-#elif defined(__linux__)
-    // Hard limit of 16 bytes including the NUL; longer names fail with ERANGE.
-    std::string const truncated = name.substr(0, 15);
-    ::pthread_setname_np(::pthread_self(), truncated.c_str());
-#elif defined(__FreeBSD__) || defined(__OpenBSD__)
-    ::pthread_set_name_np(::pthread_self(), name.c_str());
-#else
-    (void)name;
-#endif
-}
-
+void setCurrentThreadName(std::string const& name);
 #if !defined(_WIN32)
-// pthread_attr_setstacksize requires a multiple of the page size and at least
-// PTHREAD_STACK_MIN. On glibc >= 2.34 PTHREAD_STACK_MIN is no longer a
-// compile-time constant on every architecture, so prefer sysconf.
-inline std::size_t
-roundStackSize(std::size_t bytes)
-{
-    long const pageRaw = ::sysconf(_SC_PAGESIZE);
-    std::size_t const page =
-        (pageRaw > 0) ? static_cast<std::size_t>(pageRaw) : 4096u;
-
-    std::size_t minStack = page;
-#if defined(_SC_THREAD_STACK_MIN)
-    long const minRaw = ::sysconf(_SC_THREAD_STACK_MIN);
-    if (minRaw > 0)
-    {
-        minStack = std::max(minStack, static_cast<std::size_t>(minRaw));
-    }
-#endif
-#if defined(PTHREAD_STACK_MIN)
-    minStack = std::max(minStack, static_cast<std::size_t>(PTHREAD_STACK_MIN));
+std::size_t roundStackSize(std::size_t bytes);
+bool isNullThread(::pthread_t const& t) noexcept;
 #endif
 
-    bytes = std::max(bytes, minStack);
-    // Round up to a page multiple, guarding against overflow.
-    if (bytes > (~static_cast<std::size_t>(0)) - (page - 1))
-    {
-        return bytes - (bytes % page);
-    }
-    return ((bytes + page - 1) / page) * page;
-}
-#endif
-
-// Note: passing a C++-linkage function to pthread_create is formally
-// unspecified, but is what every real implementation (and libstdc++/libc++
-// themselves) does.
 #if defined(_WIN32)
-inline unsigned __stdcall trampoline(void* raw)
+unsigned __stdcall trampoline(void* raw);
 #else
-inline void*
-trampoline(void* raw)
+void* trampoline(void* raw);
 #endif
-{
-    std::unique_ptr<PayloadBase> payload(static_cast<PayloadBase*>(raw));
-    try
-    {
-        setCurrentThreadName(payload->mName);
-        payload->run();
-    }
-    catch (...)
-    {
-        // std::thread's contract: an exception escaping the thread function
-        // calls std::terminate. Letting it unwind into the C runtime here
-        // would be undefined behaviour, so make it explicit.
-        std::terminate();
-    }
-#if defined(_WIN32)
-    return 0u;
-#else
-    return nullptr;
-#endif
-}
 
 } // namespace detail
 
@@ -202,73 +106,23 @@ class StackThread
     class id
     {
       public:
-        id() noexcept : mNative{}, mValid(false)
-        {
-        }
+        id() noexcept;
 
-        explicit id(native_id_type native) noexcept
-            : mNative(native), mValid(true)
-        {
-        }
+        explicit id(native_id_type native) noexcept;
 
-        friend bool
-        operator==(id const& a, id const& b) noexcept
-        {
-            if (a.mValid != b.mValid)
-            {
-                return false;
-            }
-            if (!a.mValid)
-            {
-                return true;
-            }
-#if defined(_WIN32)
-            return a.mNative == b.mNative;
-#else
-            return ::pthread_equal(a.mNative, b.mNative) != 0;
-#endif
-        }
+        friend bool operator==(id const& a, id const& b) noexcept;
 
-        friend bool
-        operator!=(id const& a, id const& b) noexcept
-        {
-            return !(a == b);
-        }
+        friend bool operator!=(id const& a, id const& b) noexcept;
 
         // Ordering exists so `id` can key a std::map. POSIX gives no ordering
         // over pthread_t, so this compares the object representation. That is
         // a strict weak ordering consistent with == on every implementation
         // where pthread_t is a scalar (all of glibc, musl, macOS).
-        friend bool
-        operator<(id const& a, id const& b) noexcept
-        {
-            if (a.mValid != b.mValid)
-            {
-                return !a.mValid;
-            }
-            if (!a.mValid)
-            {
-                return false;
-            }
-            return std::memcmp(&a.mNative, &b.mNative, sizeof(native_id_type)) <
-                   0;
-        }
+        friend bool operator<(id const& a, id const& b) noexcept;
 
-        friend bool
-        operator>(id const& a, id const& b) noexcept
-        {
-            return b < a;
-        }
-        friend bool
-        operator<=(id const& a, id const& b) noexcept
-        {
-            return !(b < a);
-        }
-        friend bool
-        operator>=(id const& a, id const& b) noexcept
-        {
-            return !(a < b);
-        }
+        friend bool operator>(id const& a, id const& b) noexcept;
+        friend bool operator<=(id const& a, id const& b) noexcept;
+        friend bool operator>=(id const& a, id const& b) noexcept;
 
         template <class Char, class Traits>
         friend std::basic_ostream<Char, Traits>&
@@ -284,18 +138,7 @@ class StackThread
             return os << scalar;
         }
 
-        std::size_t
-        hash() const noexcept
-        {
-            if (!mValid)
-            {
-                return 0;
-            }
-            std::uintptr_t scalar = 0;
-            std::memcpy(&scalar, &mNative,
-                        std::min(sizeof(scalar), sizeof(native_id_type)));
-            return std::hash<std::uintptr_t>{}(scalar);
-        }
+        std::size_t hash() const noexcept;
 
       private:
         native_id_type mNative;
@@ -345,110 +188,19 @@ class StackThread
     StackThread(StackThread const&) = delete;
     StackThread& operator=(StackThread const&) = delete;
 
-    StackThread(StackThread&& other) noexcept
-    {
-        swap(other);
-    }
+    StackThread(StackThread&& other) noexcept;
 
-    StackThread&
-    operator=(StackThread&& other) noexcept
-    {
-        if (this != &other)
-        {
-            if (joinable())
-            {
-                // Matching std::thread: assigning over a joinable thread is
-                // a programming error, not a silent detach.
-                std::terminate();
-            }
-            swap(other);
-        }
-        return *this;
-    }
+    StackThread& operator=(StackThread&& other) noexcept;
 
-    ~StackThread()
-    {
-        if (joinable())
-        {
-            std::terminate();
-        }
-    }
+    ~StackThread();
 
-    bool
-    joinable() const noexcept
-    {
-#if defined(_WIN32)
-        return mHandle != nullptr;
-#else
-        return mJoinable;
-#endif
-    }
+    bool joinable() const noexcept;
 
-    void
-    join()
-    {
-        if (!joinable())
-        {
-            throw std::system_error(
-                std::make_error_code(std::errc::invalid_argument),
-                "StackThread::join on a non-joinable thread");
-        }
-#if defined(_WIN32)
-        if (mId == ::GetCurrentThreadId())
-        {
-            throw std::system_error(
-                std::make_error_code(std::errc::resource_deadlock_would_occur),
-                "StackThread::join on itself");
-        }
-        DWORD const rc = ::WaitForSingleObject(mHandle, INFINITE);
-        if (rc != WAIT_OBJECT_0)
-        {
-            throw std::system_error(static_cast<int>(::GetLastError()),
-                                    std::system_category(),
-                                    "WaitForSingleObject");
-        }
-        ::CloseHandle(mHandle);
-        mHandle = nullptr;
-#else
-        int const rc = ::pthread_join(mThread, nullptr);
-        if (rc != 0)
-        {
-            throw std::system_error(rc, std::system_category(), "pthread_join");
-        }
-        mJoinable = false;
-#endif
-        mId = id{};
-    }
+    void join();
 
-    void
-    detach()
-    {
-        if (!joinable())
-        {
-            throw std::system_error(
-                std::make_error_code(std::errc::invalid_argument),
-                "StackThread::detach on a non-joinable thread");
-        }
-#if defined(_WIN32)
-        ::CloseHandle(mHandle);
-        mHandle = nullptr;
-#else
-        int const rc = ::pthread_detach(mThread);
-        if (rc != 0)
-        {
-            throw std::system_error(rc, std::system_category(),
-                                    "pthread_detach");
-        }
-        mJoinable = false;
-#endif
-        mId = id{};
-    }
+    void detach();
 
-    id
-    get_id() const noexcept
-    {
-        return mId;
-    }
+    id get_id() const noexcept;
 
     // Valid only while joinable(). On POSIX this is the pthread_t, suitable for
     // pthread_setaffinity_np, pthread_setschedparam, pthread_getattr_np, etc.
@@ -460,123 +212,28 @@ class StackThread
     // a thread, do it promptly after construction, or have the thread do it to
     // itself. (Windows HANDLEs do not have this problem -- they stay queryable
     // after exit.)
-    native_handle_type
-    native_handle() const noexcept
-    {
-#if defined(_WIN32)
-        return mHandle;
-#else
-        return mThread;
-#endif
-    }
+    native_handle_type native_handle() const noexcept;
 
-    void
-    swap(StackThread& other) noexcept
-    {
-#if defined(_WIN32)
-        std::swap(mHandle, other.mHandle);
-#else
-        std::swap(mThread, other.mThread);
-        std::swap(mJoinable, other.mJoinable);
-#endif
-        std::swap(mId, other.mId);
-    }
+    void swap(StackThread& other) noexcept;
 
-    static unsigned int
-    hardware_concurrency() noexcept
-    {
-        return std::thread::hardware_concurrency();
-    }
+    static unsigned int hardware_concurrency() noexcept;
 
     // Convenience: name the calling thread. Note macOS can only name itself,
     // so there is intentionally no name-another-thread entry point.
-    static void
-    setCurrentName(std::string const& name)
-    {
-        detail::setCurrentThreadName(name);
-    }
+    static void setCurrentName(std::string const& name);
 
   private:
-    void
-    start(std::size_t stackBytes, std::unique_ptr<detail::PayloadBase> payload)
-    {
-#if defined(_WIN32)
-        if (stackBytes > static_cast<std::size_t>(UINT_MAX))
-        {
-            throw std::system_error(
-                std::make_error_code(std::errc::invalid_argument),
-                "requested stack size exceeds the Win32 unsigned limit");
-        }
-        // Without STACK_SIZE_PARAM_IS_A_RESERVATION the size argument is the
-        // initial *commit*, and the reserve still comes from the PE header.
-        unsigned threadId = 0;
-        uintptr_t const h = ::_beginthreadex(
-            nullptr, static_cast<unsigned>(stackBytes), &detail::trampoline,
-            payload.get(), STACK_SIZE_PARAM_IS_A_RESERVATION, &threadId);
-        if (h == 0)
-        {
-            throw std::system_error(errno, std::generic_category(),
-                                    "_beginthreadex");
-        }
-        payload.release(); // ownership transferred to the new thread
-        mHandle = reinterpret_cast<HANDLE>(h);
-        mId = id{static_cast<DWORD>(threadId)};
-#else
-        ::pthread_attr_t attr;
-        int rc = ::pthread_attr_init(&attr);
-        if (rc != 0)
-        {
-            throw std::system_error(rc, std::system_category(),
-                                    "pthread_attr_init");
-        }
-        struct AttrGuard
-        {
-            ::pthread_attr_t* a;
-            ~AttrGuard()
-            {
-                ::pthread_attr_destroy(a);
-            }
-        } guard{&attr};
-
-        if (stackBytes != 0)
-        {
-            rc = ::pthread_attr_setstacksize(
-                &attr, detail::roundStackSize(stackBytes));
-            if (rc != 0)
-            {
-                throw std::system_error(rc, std::system_category(),
-                                        "pthread_attr_setstacksize");
-            }
-        }
-
-        ::pthread_t tid{};
-        rc = ::pthread_create(&tid, &attr, &detail::trampoline, payload.get());
-        if (rc != 0)
-        {
-            throw std::system_error(rc, std::system_category(),
-                                    "pthread_create");
-        }
-        payload.release(); // ownership transferred to the new thread
-        mThread = tid;
-        mJoinable = true;
-        mId = id{tid};
-#endif
-    }
+    void start(std::size_t stackBytes,
+               std::unique_ptr<detail::PayloadBase> payload);
 
 #if defined(_WIN32)
     HANDLE mHandle = nullptr;
 #else
     ::pthread_t mThread{};
-    bool mJoinable = false;
 #endif
-    id mId{};
 };
 
-inline void
-swap(StackThread& a, StackThread& b) noexcept
-{
-    a.swap(b);
-}
+void swap(StackThread& a, StackThread& b) noexcept;
 
 } // namespace stellar
 
@@ -584,10 +241,6 @@ namespace std
 {
 template <> struct hash<::stellar::StackThread::id>
 {
-    std::size_t
-    operator()(::stellar::StackThread::id const& v) const noexcept
-    {
-        return v.hash();
-    }
+    std::size_t operator()(::stellar::StackThread::id const& v) const noexcept;
 };
 } // namespace std
